split api from main flask app utilize fastapi move web dashboard files to /web
