+++ web/templates/editor.html
@@ -5,7 +5,7 @@
 {% endblock %}
 {% block content %}
 <div class="container">
-  <p>Modify individual YAML config files or Markdown files for the monitoring system. Keep in mind that a restart of the service will be needed to make-live any changes. For help with syntax refer to <a href="/guide">the guide</a>.</p>
+  <p>Modify individual YAML config files or Markdown files for the monitoring system. Keep in mind that a restart of the service will be needed to make-live any changes. For help with syntax refer to <a href="/web/guide">the guide</a>.</p>
   <div align="right">
     {% if not editor_config['read_only']: %}
     <button class="btn btn-success mr-1" onClick="saveFile()">Save</button>
