+++ web/templates/performance_data.html
@@ -124,7 +124,7 @@ function load_graph(){
           <a href="?minutes=1440" class="btn dropdown-item">24 hours</a>
         </div>
       </div>
-      <a href="/status/host/{{ service['host']['id'] }}" class="btn btn-primary host_back_btn"> Host Status</a>
+      <a href="/web/status/host/{{ service['host']['id'] }}" class="btn btn-primary host_back_btn"> Host Status</a>
     </div>
   </div>
 </div>
