+++ web/templates/services.html
@@ -49,7 +49,7 @@ function load_status(){
       if(s.hasOwnProperty('perf_data'))
       {
         link2 = $('<a>');
-        link2.attr('href', '/perf_data/' + s.id);
+        link2.attr('href', '/web/perf_data/' + s.id);
         link2.attr('class', 'mdi mdi-chart-areaspline-variant icon-sm icon-inline');
         col_size = 6;
       }
@@ -71,7 +71,7 @@ function load_status(){
 
       // link the host name
       host = $('<a>').html(s.host.name);
-      host.attr('href', '/status/host/' + s.host.id);
+      host.attr('href', '/web/status/host/' + s.host.id);
 
       row.append(td.clone().append(host));
