+++ web/templates/view_tags.html
@@ -5,7 +5,7 @@
   <ul>
     {% for t in tags: %}
     <li>
-      <a href="/status/tag/{{ t }} ">{{ tags[t]['name'] }}</a>
+      <a href="/web/status/tag/{{ t }} ">{{ tags[t]['name'] }}</a>
     </li>
     {% endfor %}
   </ul>
